feat: introduce new shadcn-based toast components
