+++ packages/design-system/src/Toast.stories.tsx
@@ -1,152 +1,170 @@
-import { useState } from 'react'
+import { faCheckCircle } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import Button from './Button'
-import { H4 } from './Header'
-import Toast from './Toast'
+import { ToastAction, Toaster, useToast } from './Toast'
 
 export const Default = () => {
+  const { toast } = useToast({})
+
   return (
-    <div>
-      <div>
-        Toasts can be used with an internally managed state when a trigger can
-        be used to trigger the toast (i.e. a button).
-      </div>
-      <Toast
-        triggerText="Trigger Text"
-        title="Title"
-        description="Description"
-        duration={5000}
-      />
-    </div>
+    <>
+      Default toasts can be triggered with a title and description.
+      <Toaster />
+      <Button
+        onClick={() =>
+          toast({
+            title: 'Title',
+            description:
+              'Dolore incididunt reprehenderit officia et occaecat non laboris.',
+          })
+        }
+      >
+        Trigger
+      </Button>
+    </>
+  )
+}
+
+export const Simple = () => {
+  const { toast } = useToast({})
+
+  return (
+    <>
+      Simple toasts can be triggered with a description as content only.
+      <Toaster />
+      <Button
+        onClick={() =>
+          toast({
+            description:
+              'Dolore incididunt reprehenderit officia et occaecat non laboris.',
+          })
+        }
+      >
+        Trigger
+      </Button>
+    </>
   )
 }
 
 export const Children = () => {
-  return (
-    <div>
-      <div>
-        Toasts can be used with an internally managed state when a trigger can
-        be used to trigger the toast (i.e. a button) and children as content.
-      </div>
-      <Toast triggerText="Trigger Text" duration={5000}>
-        Content as Children
-      </Toast>
-    </div>
-  )
-}
-
-export const ExternalState = () => {
-  const [open, setOpen] = useState(false)
+  const { toast } = useToast()
 
   return (
-    <div>
-      <div>
-        Toasts can be used with an externally managed state and either
-        title/description or children as content. As both the open state and the
-        state change method have to be passed to the component, the toast will
-        still disappear after the specified duration, given that the user does
-        not hover on it.
-      </div>
-      <Button onClick={() => setOpen(true)}>Trigger now is a Button</Button>
-      <Toast
-        duration={5000}
-        openExternal={open}
-        onCloseExternal={() => setOpen(false)}
+    <>
+      Simple toasts can be triggered with a description as content only.
+      <Toaster />
+      <Button
+        onClick={() =>
+          toast({
+            children: (
+              <div className="flex flex-row items-center gap-3">
+                <FontAwesomeIcon
+                  className="text-green-600"
+                  icon={faCheckCircle}
+                />
+                <span>SUCCESS</span>
+              </div>
+            ),
+          })
+        }
       >
-        Content as Children
-      </Toast>
-    </div>
+        Trigger
+      </Button>
+    </>
   )
 }
 
 export const Action = () => {
+  const { toast } = useToast()
+
   return (
-    <div>
-      <div>
-        Toasts can make use of a radix action button (basic button version).
-        This button will be shown, if a corresponding text and onClick function
-        are provided
-      </div>
-      <Toast
-        triggerText="Trigger Text"
-        duration={5000}
-        actionText="Action Trigger"
-        actionOnClick={() => alert('Action clicked')}
+    <>
+      Toasts can have an action button that can be triggered.
+      <Toaster />
+      <Button
+        onClick={() =>
+          toast({
+            description:
+              'Dolore incididunt reprehenderit officia et occaecat non laboris.',
+            action: (
+              <ToastAction
+                altText="Action"
+                onClick={() => alert('Action clicked!')}
+              >
+                Action
+              </ToastAction>
+            ),
+          })
+        }
       >
-        Content
-      </Toast>
-    </div>
+        Trigger
+      </Button>
+    </>
   )
 }
 
 export const Success = () => {
+  const { toast } = useToast({ type: 'success' })
+
   return (
-    <div>
-      <div>Styled toast component with success type.</div>
-      <Toast triggerText="Trigger" duration={5000} type="success">
-        <H4>Title</H4>
-        <div>Content Success</div>
-      </Toast>
-    </div>
+    <>
+      Prestyled toasts can be triggered with a title and description.
+      <Toaster />
+      <Button
+        onClick={() =>
+          toast({
+            title: 'Title',
+            description:
+              'Dolore incididunt reprehenderit officia et occaecat non laboris.',
+          })
+        }
+      >
+        Trigger
+      </Button>
+    </>
   )
 }
 
 export const Warning = () => {
+  const { toast } = useToast({ type: 'warning' })
+
   return (
-    <div>
-      <div>Styled toast component with warning type.</div>
-      <Toast triggerText="Trigger" type="warning">
-        Content of Tooltip
-      </Toast>
-    </div>
+    <>
+      Prestyled toasts can be triggered with a title and description.
+      <Toaster />
+      <Button
+        onClick={() =>
+          toast({
+            title: 'Title',
+            description:
+              'Dolore incididunt reprehenderit officia et occaecat non laboris.',
+          })
+        }
+      >
+        Trigger
+      </Button>
+    </>
   )
 }
 
 export const Error = () => {
-  return (
-    <div>
-      <div>Styled toast component with error type.</div>
-      <Toast triggerText="Trigger" duration={5000} type="error">
-        Content with a bit more text so that it will be split onto multiple
-        lines due to the maximum width specified for the tooltip
-      </Toast>
-    </div>
-  )
-}
+  const { toast } = useToast({ type: 'error' })
 
-export const Dismissible = () => {
   return (
-    <div className="flex flex-col gap-2">
-      <div>
-        Toast components can have an "addimissible" prop set to true. This will
-        set the duration automatically to a minute, but offer an option to the
-        user to manually close the tooltip. Optionally, the duration of a minute
-        can also be overwritten using the duration prop.
-      </div>
-      <Toast triggerText="Trigger - no overwritten duration" dismissible>
-        Toast Content
-      </Toast>
-      <Toast triggerText="Trigger - duration 5s" duration={5000} dismissible>
-        Toast Content
-      </Toast>
-      <Toast
-        type="success"
-        triggerText="Success Toast"
-        duration={5000}
-        dismissible
+    <>
+      Prestyled toasts can be triggered with a title and description.
+      <Toaster />
+      <Button
+        onClick={() =>
+          toast({
+            title: 'Title',
+            description:
+              'Dolore incididunt reprehenderit officia et occaecat non laboris.',
+          })
+        }
       >
-        Toast Content
-      </Toast>
-      <Toast
-        type="warning"
-        triggerText="Warning Toast"
-        duration={5000}
-        dismissible
-      >
-        Toast Content
-      </Toast>
-      <Toast type="error" triggerText="Error Toast" duration={5000} dismissible>
-        Toast Content
-      </Toast>
-    </div>
+        Trigger
+      </Button>
+    </>
   )
 }
