+++ packages/design-system/src/index.ts
@@ -30,6 +30,7 @@ export * from './TabsLegacy'
 export { default as TabsLegacy } from './TabsLegacy'
 export * from './Tag'
 export * from './Toast'
+export * from './ToastLegacy'
 export * from './Tooltip'
 export * from './UserNotification'
 export * from './Workflow'
