+++ packages/design-system/src/ui/use-toast.ts
@@ -6,7 +6,7 @@ import type { ToastActionElement, ToastProps } from '@/ui/toast'
 const TOAST_LIMIT = 1
 const TOAST_REMOVE_DELAY = 1000000
 
-type ToasterToast = ToastProps & {
+export type ToasterToast = ToastProps & {
   id: string
   title?: React.ReactNode
   description?: React.ReactNode
